refactor: adjust backend file structure for clarity
