+++ apps/backend/src/config/limiter.ts
@@ -1,0 +1,3 @@
+import rateLimit from "express-rate-limit";
+
+export const limiter = rateLimit({ windowMs: 15 * 60 * 1000, max: 100 });
