+++ apps/backend/src/index.ts
@@ -1,4 +1,4 @@
-import { env } from "./env";
+import { env } from "./config/env";
 import { createServer } from "./server";
 
 const app = createServer();
