+++ apps/backend/src/middleware/isLoggedIn.ts
@@ -1,0 +1,8 @@
+import { RequestHandler } from "express";
+
+export const isLoggedIn: RequestHandler = (req, res, next) => {
+    if (req.user) {
+        return next();
+    }
+    res.status(401).json({ message: "Unauthorized" });
+};
