+++ apps/backend/src/routes/logout.route.ts
@@ -1,0 +1,6 @@
+import { Router } from "express";
+import { logout } from "../controllers/logout.controller";
+
+export const logoutRouter: Router = Router();
+
+logoutRouter.get("/", logout);
